refactor: move ineffective dynamic import detection into pre_rendered_chunk generation
